pyproject.toml
[tool.poetry]
name = "hyperion"
version = "0.1.0.dev1"
description = "Next Generation Architecture for GVM"
authors = ["Greenbone Networks <info@greenbone.net>"]
readme = "README.md"
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
graphene-django = "^2.9.0"
python-gvm = "^21.1.1"
django-cors-headers = "^3.2.1"


[tool.poetry.dev-dependencies]
autohooks = "^2.2.0"
autohooks-plugin-black = "^1.2.0"
autohooks-plugin-pylint = "^1.2.0"
pylint-django = "^2.3.0"
coverage = "^5.0.4"
sphinx = "^2.4.4"
black = "20.8b1"
pontos = "^0.3.0"
icalendar = "^4.0.6"

[tool.black]
line-length = 80
target-version = ['py36', 'py37', 'py38']
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.hg
  | \.venv
  | \.circleci
  | \.github
  | \.vscode
  | _build
  | build
  | dist
  | docs
)/
'''

[tool.autohooks]
mode = "poetry"
pre-commit = ['autohooks.plugins.black', 'autohooks.plugins.pylint']

[tool.pontos.version]
version-module-file = "hyperion/__version__.py"
